requirements.txt
# AWS Cost Optimizer for Interns - Dependencies
# Core dependencies for the cost optimization system

# Groq API (OpenAI-compatible chat completions)
groq>=0.11.0

# Environment variable management
python-dotenv>=1.0.0

# JSON handling (built-in, but ensuring compatibility)
# json - built-in module

# HTTP requests for API calls (if needed for future enhancements)
requests>=2.31.0

# Data manipulation and analysis
pandas>=2.0.0

# Date and time handling
python-dateutil>=2.8.0

# Optional: For enhanced CLI experience
rich>=13.0.0
click>=8.1.0

# Optional: For data visualization (future enhancement)
matplotlib>=3.7.0
seaborn>=0.12.0

# Enterprise dashboard
streamlit>=1.32.0
plotly>=5.18.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
